Add files via upload
This summer school project is called arithmetic.  It is given a order of operations (L or N) and a product/sum and an arbitrary amount of positive integers.  The program then uses a stack which works as a depth first search tree and assigns operators + or * in-between each number until it comes to:
Solution or Not possible.
Left to Right evaluation using addition and multiplication. (L)
Left to Right evaluation using addition and multiplication - but multiplication takes precedence.  (N)